Use unsigned int instead of unsigned long for ID and UID.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@4232 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qmimap4/src/buffer.h
@@ -118,7 +118,7 @@ struct TokenUtil
 	static bool isEqualIgnoreCase(const std::pair<const CHAR*, size_t>& s,
 								  const CHAR* psz);
 	static bool string2number(const std::pair<const CHAR*, size_t>& s,
-							  unsigned long* pn);
+							  unsigned int* pn);
 };
 
 }
